Update view2
so for this one I got an empty output for the table, couldn't figure out why. When I ran the innermost subquery, for some reason the VisitID would print for every prescription name BUT "Vicodin". I even changed the names around to check but the VisitID would NOT print where Vicodin was the prescription name.